New Lighting algorithm for Standard Shaders (need to backport to OpenGL ones)

+++ C3DE.Content/Shaders/Forward/OpenGL/StandardBase.fx
@@ -18,7 +18,7 @@ float4x4 Projection;
 
 // Material
 float3 AmbientColor;
-float SpecularPower;
+int SpecularPower;
 
 // Misc
 float3 EyePosition;
